# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:1400",
  "basePath": "/",
  "tags": [
    {
      "name": "User",
      "description": "Endpoints"
    },
    {
      "name": "Category",
      "description": "Category Endpoints"
    },
    {
      "name": "Donation",
      "description": "Donation Endpoints"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-KEY",
      "description": "any description..."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/register": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to sign in a specific user",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddUser"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to sign in a specific user",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/categories": {
      "get": {
        "tags": [
          "Category"
        ],
        "description": "Endpoint to fetch a categories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/Category": {
      "post": {
        "tags": [
          "Category"
        ],
        "description": "Endpoint to add a category",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Category information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddCategory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/category/{id}": {
      "get": {
        "tags": [
          "Category"
        ],
        "description": "Endpoint to get a category",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/category/{id}/deactivate": {
      "put": {
        "tags": [
          "Category"
        ],
        "description": "Endpoint to delete a category",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/category/{id}/edit": {
      "put": {
        "tags": [
          "Category"
        ],
        "description": "Endpoint to edit a category",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "Category information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateCategory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/category/{id}/activate": {
      "put": {
        "tags": [
          "Category"
        ],
        "description": "Endpoint to activate a category",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/donations": {
      "get": {
        "tags": [
          "Donation"
        ],
        "description": "Endpoint to fetch a categories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/donation": {
      "post": {
        "tags": [
          "Donation"
        ],
        "description": "Endpoint to add a Donation",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Donation information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddDonation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/donation/{id}": {
      "get": {
        "tags": [
          "Donation"
        ],
        "description": "Endpoint to get a Donation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/donation/{id}/deactivate": {
      "put": {
        "tags": [
          "Donation"
        ],
        "description": "Endpoint to delete a donation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/donation/{id}/edit": {
      "put": {
        "tags": [
          "Donation"
        ],
        "description": "Endpoint to edit a donation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "Donation information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateDonation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/donation/{id}/activate": {
      "put": {
        "tags": [
          "Donation"
        ],
        "description": "Endpoint to activate a Donation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    }
  },
  "definitions": {
    "AddUser": {
      "type": "object",
      "properties": {
        "first_name": {
          "type": "string",
          "example": "John"
        },
        "surname": {
          "type": "string",
          "example": "Doe"
        },
        "email": {
          "type": "string",
          "example": "email@example.com"
        },
        "phone": {
          "type": "string",
          "example": "07279797819"
        },
        "password": {
          "type": "string",
          "example": "passwordexample#com"
        },
        "password_confirm": {
          "type": "string",
          "example": "passwordexample#com"
        }
      },
      "required": [
        "first_name",
        "surname"
      ]
    },
    "LoginUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "email@example.com"
        },
        "password": {
          "type": "string",
          "example": "passwordexample#com"
        }
      }
    },
    "AddCategory": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "sample category"
        }
      }
    },
    "EditCategory": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "sample category"
        }
      }
    },
    "AddDonation": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "sample Donation"
        },
        "category_id": {
          "type": "string",
          "example": "60406e9a5a75fd34144f7451"
        }
      }
    },
    "EditDonation": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "sample Donation"
        },
        "category_id": {
          "type": "string",
          "example": "60406e9a5a75fd34144f7451"
        }
      }
    }
  }
}